% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polity.R
\docType{data}
\name{polity}
\alias{polity}
\title{Polity 5 Snippet}
\format{
An object of class \code{data.frame}
}
\usage{
data(polity)
}
\description{
Data from the Polity 5 dataset
}
\references{
Marshall, Monty G., Ted Robert Gurr. 2020. "Polity5: Political
Regime Characteristics and Transitions, 1800-2018". Dataset Users' Manual.
Center for Systemic Peace. Available at http://www.systemicpeace.org/inscr/p5manualv2018.pdf
}
\keyword{datasets}
